Gloal Exception CURD author, category, pov

+++ Backend/Novelbackend/src/main/java/com/example/demo/service/AuthorService.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Service;
 
 import com.example.demo.dto.request.AuthorCreateionRequest;
 import com.example.demo.dto.request.AuthorUpdateRequest;
-import com.example.demo.dto.request.CategoryUpdateRequest;
 import com.example.demo.dto.respone.AuthorRespone;
-import com.example.demo.dto.respone.CategoryRespone;
 import com.example.demo.entity.Author;
+import com.example.demo.exception.AppException;
+import com.example.demo.exception.ErrorCode;
 import com.example.demo.mapper.IAuthorMapper;
 import com.example.demo.repository.IAuthorRepository;
 import lombok.AccessLevel;
@@ -35,12 +35,17 @@ public class AuthorService {
 		return authorRepository.findAll().stream().map(t -> authorMapper.toAuthorRespone(t)).toList();
 	}
 	public String deleteAuthor(String idAuthor) {
+		if (!authorRepository.existsById(idAuthor)) {
+			throw new AppException(ErrorCode.AUTHOR_NOT_EXISTED);
+		}
 		authorRepository.deleteById(idAuthor);
 		return idAuthor;
 	}
 
 	public Optional<AuthorRespone> updateAuthor(AuthorUpdateRequest request) {
-		
+		if (!authorRepository.existsById(request.getIdAuthor())) {
+			throw new AppException(ErrorCode.AUTHOR_NOT_EXISTED);
+		}
 		return authorRepository.findById(request.getIdAuthor()).map(t -> {
 			t.setDescriptionAuthor(request.getDescriptionAuthor());
 			t.setNameAuthor(t.getNameAuthor());
